// apps/web/src/lib/api/generated/model/planOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */
import type { PlanOutLimits } from './planOutLimits';
import type { PlanPrice } from './planPrice';

export interface PlanOut {
  code: string;
  extra_seat_monthly: string | null;
  features: string[];
  included_seats: number;
  limits: PlanOutLimits;
  name: string;
  prices: PlanPrice[];
  tagline: string;
}