// apps/web/src/lib/api/generated/model/subscriptionPaymentOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export interface SubscriptionPaymentOut {
  /** @pattern ^-?\d{1,16}(\.\d{1,4})?$ */
  amount: string;
  billing_cycle: string;
  completed_at: string | null;
  created_at: string;
  discount_percent: number;
  extra_seats: number;
  id: string;
  /** @pattern ^-?\d{1,16}(\.\d{1,4})?$ */
  list_price: string;
  period_end: string | null;
  period_start: string | null;
  plan: string;
  receipt: string | null;
  result_desc: string | null;
  /** pending, paid, cancelled, failed or expired */
  status: string;
}